Backport: feature(lu600): fan controller and enabled all I2C
Author: Hitesh Viradiya <Hitesh@liveu.tv>
Date:   Thu Dec 3 20:38:47 2015 +0530

    feature(lu600): fan controller and enabled all I2C

    This patch adds Fan controller support as like Sitara core card and also enables
    all the I2Cs.

    Change-Id: Id06bdb20c3f97fadf40c878dd562b764f56c61a5
    Signed-off-by: Hitesh Viradiya <Hitesh@liveu.tv>

In addition to that removed unwanted I2C initialization. And defconfig
mechanism has been changed according to that all the config (CONFIG_*)
must be provided through Kconfig file. In config (CONFIG_*) is already
present in Kconfig file then it can be enabled via "menuconfig" or to
make this config permenantly enabled it can be added in board specific
defconfig file (in our case liveuboard_defconfig).

Signed-off-by: Jemish Patel <jemish@liveu.tv>